test: split secretscan feedback coverage
