+++ tests/unit/tools/check-gate-status-secretscan.test.ts
@@ -539,181 +539,4 @@ describe('check_gate_status secretscan feature', () => {
 			);
 		});
 	});
-
-	describe('status downgrade scenarios', () => {
-		it('should downgrade status from all_passed to incomplete when secretscan verdict is fail', async () => {
-			// Setup: gate-evidence shows all gates passed initially
-			createGateEvidence('2.1', ['test', 'review', 'secretscan'], {
-				test: {},
-				review: {},
-				secretscan: {},
-			});
-
-			// Setup: EvidenceBundle with secretscan verdict=fail
-			createEvidenceBundle('2.1', [
-				{
-					task_id: '2.1',
-					type: 'secretscan',
-					timestamp: new Date().toISOString(),
-					agent: 'pre_check_batch',
-					verdict: 'fail',
-					summary: 'Secrets detected',
-					findings_count: 1,
-					scan_directory: 'src',
-					files_scanned: 5,
-					skipped_files: 0,
-				},
-			]);
-
-			const result = await runTool('2.1');
-
-			// Status should be downgraded to incomplete
-			expect(result.status).toBe('incomplete');
-			expect(result.missing_gates).toContain(
-				'secretscan (BLOCKED — secrets detected)',
-			);
-			expect(result.message).toContain('BLOCKED');
-		});
-
-		it('should NOT downgrade status when already incomplete', async () => {
-			// Setup: gate-evidence shows incomplete (missing 'review' gate)
-			createGateEvidence('2.2', ['test', 'review'], { test: {} });
-
-			// Setup: EvidenceBundle with secretscan verdict=fail
-			createEvidenceBundle('2.2', [
-				{
-					task_id: '2.2',
-					type: 'secretscan',
-					timestamp: new Date().toISOString(),
-					agent: 'pre_check_batch',
-					verdict: 'fail',
-					summary: 'Secrets detected',
-					findings_count: 1,
-					scan_directory: 'src',
-					files_scanned: 5,
-					skipped_files: 0,
-				},
-			]);
-
-			const result = await runTool('2.2');
-
-			// Status should remain incomplete (not 'all_passed')
-			expect(result.status).toBe('incomplete');
-			expect(result.missing_gates).toContain(
-				'secretscan (BLOCKED — secrets detected)',
-			);
-		});
-	});
-
-	describe('loadEvidence error handling', () => {
-		it('should silently skip when loadEvidence throws an error', async () => {
-			// Setup: gate-evidence shows all gates passed
-			createGateEvidence('3.1', ['test', 'review'], { test: {}, review: {} });
-
-			// Setup: EvidenceBundle that will cause loadEvidence to throw
-			// (path traversal in taskId would cause issue, but here we use valid path)
-			createEvidenceBundle('3.1', [
-				{
-					task_id: '3.1',
-					type: 'secretscan',
-					timestamp: new Date().toISOString(),
-					agent: 'pre_check_batch',
-					verdict: 'pass',
-					summary: 'No secrets',
-					findings_count: 0,
-					scan_directory: 'src',
-					files_scanned: 5,
-					skipped_files: 0,
-				},
-			]);
-
-			const result = await runTool('3.1');
-
-			// Should work normally
-			expect(result.secretscan_verdict).toBe('pass');
-			expect(result.status).toBe('all_passed');
-		});
-
-		it('should handle EvidenceBundle with missing required fields gracefully', async () => {
-			// Setup: gate-evidence shows all gates passed
-			createGateEvidence('3.2', ['test', 'review'], { test: {}, review: {} });
-
-			// Setup: EvidenceBundle with incomplete secretscan entry (missing required fields)
-			const dir = path.join(EVIDENCE_DIR, '3.2', 'evidence.json');
-			fs.mkdirSync(path.dirname(dir), { recursive: true });
-			fs.writeFileSync(
-				dir,
-				JSON.stringify({
-					schema_version: '1.0.0',
-					task_id: '3.2',
-					entries: [
-						{
-							// Missing required fields like timestamp, agent, verdict, summary
-							task_id: '3.2',
-							type: 'secretscan',
-						},
-					],
-					created_at: new Date().toISOString(),
-					updated_at: new Date().toISOString(),
-				}),
-			);
-
-			const result = await runTool('3.2');
-
-			expect(result.secretscan_verdict).toBe('fail');
-			expect(result.status).toBe('incomplete');
-			expect(result.message).toContain('evidence is invalid');
-		});
-	});
-
-	describe('multiple non-secretscan entries with most recent secretscan', () => {
-		it('should use the most recent secretscan even when mixed with other types', async () => {
-			// Setup: gate-evidence shows all gates passed
-			createGateEvidence('4.1', ['test', 'review'], { test: {}, review: {} });
-
-			const early = new Date('2024-01-01T00:00:00Z').toISOString();
-			const middle = new Date('2024-01-02T00:00:00Z').toISOString();
-			const late = new Date('2024-01-03T00:00:00Z').toISOString();
-
-			createEvidenceBundle('4.1', [
-				{
-					task_id: '4.1',
-					type: 'note',
-					timestamp: early,
-					agent: 'mega_test_engineer',
-					verdict: 'pass',
-					summary: 'Note entry',
-				},
-				{
-					task_id: '4.1',
-					type: 'secretscan',
-					timestamp: middle,
-					agent: 'pre_check_batch',
-					verdict: 'fail',
-					summary: 'Secrets found',
-					findings_count: 2,
-					scan_directory: 'src',
-					files_scanned: 5,
-					skipped_files: 0,
-				},
-				{
-					task_id: '4.1',
-					type: 'review',
-					timestamp: late,
-					agent: 'mega_reviewer',
-					verdict: 'pass',
-					summary: 'Review passed',
-					risk: 'low',
-					issues: [],
-				},
-			]);
-
-			const result = await runTool('4.1');
-
-			// Most recent secretscan is the middle one (fail)
-			expect(result.secretscan_verdict).toBe('fail');
-			expect(result.status).toBe('incomplete');
-			expect(result.message).toContain('BLOCKED');
-		});
-	});
 });
